Monitoring platform events with zipkin/opentracing
Sometimes, we need lot more than latency of a span to find out any
bottlenecks, performance issues etc in a trace. Platform metrics or perf
events provide much more insights into the performance of a span or a
sequence of spans in a trace. When we talk about platform metrics, they
are huge in number and are incoherent. This patch includes a library
called "perfevents" which can be simply imported and used. This library
is an abstraction of the platform side of counters and operations. Right
now, only the generic perf events are supported. With this patch, one
should be able to find out the cpu-cycles, instructions,
cache-references, cache-misses, branch-instructions, branch-misses and
bus cycles for a span.

The required event should be supplied in the form of a string to the
StartSpan call and then, the opentracing-go library and the
zipkin-go-opentracing library should take over from there. As soon as
the span is closed, the perf event count is logged with span.Log().

Signed-off-by: Hemant Kumar <hemant@linux.vnet.ibm.com>